Add capability to accept any SSL certificate
Issue #9 : https://github.com/AmadeusITGroup/sonar-stash/issues/9

The hostname in the Stash base URL field has to be complete in order to match the server SSL certificate's one.
If an alias is used in the base URL field, the possibility to skip hostname verification is provided by the certificates acceptance field.